nth_line_desc() has been improved using code from goto_line()

+++ src/protos.h
@@ -312,7 +312,7 @@ bool ne_ispunct(const int c, const int encoding);
 bool ne_isspace(const int c, const int encoding);
 bool ne_isword(const int c, const int encoding);
 int context_prefix(const buffer *b, char **p, int64_t *prefix_pos);
-line_desc *nth_line_desc(const buffer *b, const int n);
+line_desc *nth_line_desc(const buffer *b, const int64_t n);
 
 /* undo.c */
 void start_undo_chain(buffer *b);
